+++ CHANGELOG.md
@@ -2,9 +2,15 @@
 
 ## [Unreleased]
 
+- Add portraits based on Arachne dresses by punishedmat
+
 ## [1.1.2] - 2025-11-25
 
+- bugfix
+
 ## [1.1.1] - 2025-11-24
+
+- bugfix
 
 ## [1.1.0] - 2025-11-24